requirements/base.txt
Django==4.2.*
djangorestframework==3.14.*
djangorestframework-simplejwt==5.*
drf-spectacular==0.27.*
psycopg2-binary==2.9.*
redis==5.*
celery==5.*
flower==2.*
django-celery-beat==2.*
django-redis==5.*
cryptography==42.*
python-decouple==3.*
gunicorn==21.*
whitenoise==6.*
bleach==6.*
django-cors-headers==4.*
django-ratelimit==4.*
factory-boy==3.*
Pillow==10.*
